+++ com.unity.render-pipelines.high-definition/Editor/Lighting/HDLightUI.cs
@@ -62,6 +62,7 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             ShadowQuality = 1 << 8
         }
 
+        // That's not a real English word...
         enum Advanceable
         {
             General = 1 << 0,
@@ -211,6 +212,11 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
                         SyncLightAndShadowLayers(serialized, owner);
                 }
             }
+
+            if (serialized.editorLightShape == LightShape.Directional)
+            {
+                serialized.serializedLightData.interactsWithSky.boolValue = EditorGUILayout.Toggle(s_Styles.interactsWithSky, serialized.serializedLightData.interactsWithSky.boolValue);
+            }
         }
 
         static void DrawShapeContent(SerializedHDLight serialized, Editor owner)
@@ -255,14 +261,24 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
                                 break;
                             case SpotLightShape.Cone:
                                 // Cone spot projector
+                                EditorGUI.BeginChangeCheck();
                                 EditorGUILayout.Slider(serialized.settings.spotAngle, 1f, 179f, s_Styles.outterAngle);
+                                if(EditorGUI.EndChangeCheck())
+                                {
+                                    serialized.serializedLightData.customSpotLightShadowCone.floatValue = Math.Min(serialized.serializedLightData.customSpotLightShadowCone.floatValue, serialized.settings.spotAngle.floatValue);
+                                }
                                 EditorGUILayout.Slider(serialized.serializedLightData.spotInnerPercent, 0f, 100f, s_Styles.spotInnerPercent);
                                 EditorGUILayout.PropertyField(serialized.serializedLightData.shapeRadius, s_Styles.lightRadius);
                                 EditorGUILayout.PropertyField(serialized.serializedLightData.maxSmoothness, s_Styles.maxSmoothness);
                                 break;
                             case SpotLightShape.Pyramid:
                                 // pyramid spot projector
+                                EditorGUI.BeginChangeCheck();
                                 serialized.settings.DrawSpotAngle();
+                                if (EditorGUI.EndChangeCheck())
+                                {
+                                    serialized.serializedLightData.customSpotLightShadowCone.floatValue = Math.Min(serialized.serializedLightData.customSpotLightShadowCone.floatValue, serialized.settings.spotAngle.floatValue);
+                                }
                                 EditorGUILayout.Slider(serialized.serializedLightData.aspectRatio, 0.05f, 20.0f, s_Styles.aspectRatioPyramid);
                                 EditorGUILayout.PropertyField(serialized.serializedLightData.shapeRadius, s_Styles.lightRadius);
                                 EditorGUILayout.PropertyField(serialized.serializedLightData.maxSmoothness, s_Styles.maxSmoothness);
@@ -410,7 +426,7 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
                     intensity = LightUtils.ConvertEvToLux(intensity, serialized.serializedLightData.luxAtDistance.floatValue);
             }
             else  // For area lights
-            {               
+            {
                 if (oldLightUnit == LightUnit.Lumen && newLightUnit == LightUnit.Luminance)
                     intensity = LightUtils.ConvertAreaLightLumenToLuminance((LightTypeExtent)serialized.serializedLightData.lightTypeExtent.enumValueIndex, intensity, serialized.serializedLightData.shapeWidth.floatValue, serialized.serializedLightData.shapeHeight.floatValue);
                 if (oldLightUnit == LightUnit.Luminance && newLightUnit == LightUnit.Lumen)
@@ -447,7 +463,7 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
                 }
                 else
                     EditorGUILayout.PropertyField(serialized.settings.color, s_Styles.color);
-                
+
                 if (changes.changed && HDRenderPipelinePreferences.lightColorNormalization)
                     serialized.settings.color.colorValue = HDUtils.NormalizeColor(serialized.settings.color.colorValue);
             }
@@ -527,7 +543,7 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             {
                 EditorGUILayout.ObjectField( serialized.serializedLightData.areaLightCookie, s_Styles.areaLightCookie );
             }
-            
+
             if (EditorGUI.EndChangeCheck())
             {
                 serialized.needUpdateAreaLightEmissiveMeshComponents = true;
@@ -595,12 +611,18 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
                         EditorGUILayout.PropertyField(serialized.serializedLightData.useScreenSpaceShadows, s_Styles.useScreenSpaceShadows);
                     }
 #endif
+                    if (!serialized.settings.isCompletelyBaked)
+                    {
+                        EditorGUILayout.PropertyField(serialized.serializedShadowData.shadowUpdateMode, s_Styles.shadowUpdateMode);
+                    }
+
                     using (var change = new EditorGUI.ChangeCheckScope())
                     {
                         EditorGUILayout.DelayedIntField(serialized.serializedShadowData.resolution, s_Styles.shadowResolution);
                         if (change.changed)
                             serialized.serializedShadowData.resolution.intValue = Mathf.Max(HDShadowManager.k_MinShadowMapResolution, serialized.serializedShadowData.resolution.intValue);
                     }
+
                     EditorGUILayout.Slider(serialized.serializedLightData.shadowNearPlane, HDShadowUtils.k_MinShadowNearPlane, 10f, s_Styles.shadowNearPlane);
 
                     if (serialized.settings.isMixed)
@@ -696,6 +718,18 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
                         EditorGUI.indentLevel--;
                     }
 
+                    if(serialized.editorLightShape == LightShape.Spot)
+                    {
+                        var spotLightShape = (SpotLightShape)serialized.serializedLightData.spotLightShape.enumValueIndex;
+                        if(spotLightShape != SpotLightShape.Box)
+                        {
+                            EditorGUILayout.PropertyField(serialized.serializedLightData.useCustomSpotLightShadowCone, s_Styles.useCustomSpotLightShadowCone);
+                            if(serialized.serializedLightData.useCustomSpotLightShadowCone.boolValue)
+                            {
+                                EditorGUILayout.Slider(serialized.serializedLightData.customSpotLightShadowCone, 1.0f, serialized.settings.spotAngle.floatValue, s_Styles.customSpotLightShadowCone);
+                            }
+                        }
+                    }
                 }
 
                 // Dimmer and Tint don't have effect on baked shadow