+++ com.unity.shadergraph/Editor/Data/Nodes/Input/Geometry/PositionNode.cs
@@ -1,5 +1,8 @@
+using System.Collections.Generic;
+using System.Linq;
 using UnityEngine;
 using UnityEditor.Graphing;
+using UnityEditor.ShaderGraph.Drawing.Controls;
 
 namespace UnityEditor.ShaderGraph
 {
@@ -9,7 +12,7 @@ namespace UnityEditor.ShaderGraph
     {
         private const int kOutputSlotId = 0;
         public const string kOutputSlotName = "Out";
-
+        public override List<CoordinateSpace> validSpaces => new List<CoordinateSpace> {CoordinateSpace.Object, CoordinateSpace.View, CoordinateSpace.World, CoordinateSpace.Tangent, CoordinateSpace.AbsoluteWorld};
 
         public PositionNode()
         {
@@ -28,6 +31,17 @@ namespace UnityEditor.ShaderGraph
                     SlotType.Output,
                     Vector3.zero));
             RemoveSlotsNameNotMatching(new[] { kOutputSlotId });
+        }
+
+        public override int GetCompiledNodeVersion() => 1;
+
+        public override void UpgradeNodeWithVersion(int from, int to)
+        {
+            if (from == 0 && to == 1 && space == CoordinateSpace.World)
+            {
+                var names = validSpaces.Select(cs => cs.ToString()).ToArray();
+                spacePopup = new PopupList(names, (int)CoordinateSpace.AbsoluteWorld);
+            }
         }
 
         public override string GetVariableNameForSlot(int slotId)