+++ com.unity.visualeffectgraph/Editor/Core/VFXEnums.cs
@@ -3,35 +3,35 @@ namespace UnityEditor.VFX
     // TODO Tmp Just map the task types from bindings so that the enum is accessible from outside the package (For HDRP)
     enum VFXTaskType
     {
-        None                        = UnityEngine.Experimental.VFX.VFXTaskType.None,
+        None                        = UnityEngine.VFX.VFXTaskType.None,
 
-        Spawner                     = UnityEngine.Experimental.VFX.VFXTaskType.Spawner,
-        Initialize                  = UnityEngine.Experimental.VFX.VFXTaskType.Initialize,
-        Update                      = UnityEngine.Experimental.VFX.VFXTaskType.Update,
-        Output                      = UnityEngine.Experimental.VFX.VFXTaskType.Output,
+        Spawner                     = UnityEngine.VFX.VFXTaskType.Spawner,
+        Initialize                  = UnityEngine.VFX.VFXTaskType.Initialize,
+        Update                      = UnityEngine.VFX.VFXTaskType.Update,
+        Output                      = UnityEngine.VFX.VFXTaskType.Output,
 
         // updates
-        CameraSort                  = UnityEngine.Experimental.VFX.VFXTaskType.CameraSort,
-        StripSort                   = UnityEngine.Experimental.VFX.VFXTaskType.StripSort,
-        StripUpdatePerParticle      = UnityEngine.Experimental.VFX.VFXTaskType.StripUpdatePerParticle,
-        StripUpdatePerStrip         = UnityEngine.Experimental.VFX.VFXTaskType.StripUpdatePerStrip,
+        CameraSort                  = UnityEngine.VFX.VFXTaskType.CameraSort,
+        StripSort                   = UnityEngine.VFX.VFXTaskType.StripSort,
+        StripUpdatePerParticle      = UnityEngine.VFX.VFXTaskType.StripUpdatePerParticle,
+        StripUpdatePerStrip         = UnityEngine.VFX.VFXTaskType.StripUpdatePerStrip,
 
         // outputs
-        ParticlePointOutput         = UnityEngine.Experimental.VFX.VFXTaskType.ParticlePointOutput,
-        ParticleLineOutput          = UnityEngine.Experimental.VFX.VFXTaskType.ParticleLineOutput,
-        ParticleQuadOutput          = UnityEngine.Experimental.VFX.VFXTaskType.ParticleQuadOutput,
-        ParticleHexahedronOutput    = UnityEngine.Experimental.VFX.VFXTaskType.ParticleHexahedronOutput,
-        ParticleMeshOutput          = UnityEngine.Experimental.VFX.VFXTaskType.ParticleMeshOutput,
-        ParticleTriangleOutput      = UnityEngine.Experimental.VFX.VFXTaskType.ParticleTriangleOutput,
-        ParticleOctagonOutput       = UnityEngine.Experimental.VFX.VFXTaskType.ParticleOctagonOutput,
+        ParticlePointOutput         = UnityEngine.VFX.VFXTaskType.ParticlePointOutput,
+        ParticleLineOutput          = UnityEngine.VFX.VFXTaskType.ParticleLineOutput,
+        ParticleQuadOutput          = UnityEngine.VFX.VFXTaskType.ParticleQuadOutput,
+        ParticleHexahedronOutput    = UnityEngine.VFX.VFXTaskType.ParticleHexahedronOutput,
+        ParticleMeshOutput          = UnityEngine.VFX.VFXTaskType.ParticleMeshOutput,
+        ParticleTriangleOutput      = UnityEngine.VFX.VFXTaskType.ParticleTriangleOutput,
+        ParticleOctagonOutput       = UnityEngine.VFX.VFXTaskType.ParticleOctagonOutput,
 
         // spawners
-        ConstantRateSpawner         = UnityEngine.Experimental.VFX.VFXTaskType.ConstantRateSpawner,
-        BurstSpawner                = UnityEngine.Experimental.VFX.VFXTaskType.BurstSpawner,
-        PeriodicBurstSpawner        = UnityEngine.Experimental.VFX.VFXTaskType.PeriodicBurstSpawner,
-        VariableRateSpawner         = UnityEngine.Experimental.VFX.VFXTaskType.VariableRateSpawner,
-        CustomCallbackSpawner       = UnityEngine.Experimental.VFX.VFXTaskType.CustomCallbackSpawner,
-        SetAttributeSpawner         = UnityEngine.Experimental.VFX.VFXTaskType.SetAttributeSpawner,
+        ConstantRateSpawner         = UnityEngine.VFX.VFXTaskType.ConstantRateSpawner,
+        BurstSpawner                = UnityEngine.VFX.VFXTaskType.BurstSpawner,
+        PeriodicBurstSpawner        = UnityEngine.VFX.VFXTaskType.PeriodicBurstSpawner,
+        VariableRateSpawner         = UnityEngine.VFX.VFXTaskType.VariableRateSpawner,
+        CustomCallbackSpawner       = UnityEngine.VFX.VFXTaskType.CustomCallbackSpawner,
+        SetAttributeSpawner         = UnityEngine.VFX.VFXTaskType.SetAttributeSpawner,
     }
 }