+++ com.unity.visualeffectgraph/Editor/GraphView/Elements/Controllers/VFXParameterController.cs
@@ -7,7 +7,7 @@ using UnityEditor.Experimental.GraphView;
 
 using UnityEngine;
 using UnityEngine.Profiling;
-using UnityEngine.Experimental.VFX;
+using UnityEngine.VFX;
 using UnityEngine.UIElements;
 using System.Collections.ObjectModel;