+++ com.unity.visualeffectgraph/Editor/GraphView/Views/Controller/VFXViewController.cs
@@ -5,7 +5,7 @@ using System.Collections.ObjectModel;
 using System.IO;
 using System.Linq;
 using UnityEditor.Experimental.GraphView;
-using UnityEditor.Experimental.VFX;
+using UnityEditor.VFX;
 using UnityEngine;
 using UnityEngine.Profiling;