+++ com.unity.visualeffectgraph/Editor/Inspector/VFXContextEditor.cs
@@ -4,11 +4,10 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEditor;
 using UnityEditor.Experimental;
-using UnityEditor.Experimental.VFX;
+using UnityEditor.VFX;
 using UnityEditor.SceneManagement;
 using UnityEngine;
-using UnityEngine.Experimental.VFX;
-using UnityEditor.VFX;
+using UnityEngine.VFX;
 using UnityEditor.VFX.UI;
 
 using Object = UnityEngine.Object;