+++ com.unity.shadergraph/CHANGELOG.md
@@ -9,9 +9,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - You can now use the `SHADERGRAPH_PREVIEW` keyword in `Custom Function Node` to generate different code for preview Shaders.
 - Color Mode improves node visibility by coloring the title bar by Category, Precision, or custom colors.
 - You can now set the precision of a Shader Graph and individual nodes.
+- Added _TimeParameters variable containing Time, Sin(Time), Cosine(Time)
+- **Asbolute World** space on `Position Node` now provides absolute world space coordinates regardless of the active render pipeline.
 
 ### Changed
 - The `Custom Function Node` now uses an object field to reference its source when using `File` mode.
+- To enable master nodes to generate correct motion vectors for time-based vertex modification, time is now implemented as an input to the graph rather than as a global uniform.
+- **World** space on `Position Node` now uses the default world space coordinates of the active render pipeline. 
 
 ### Fixed
 - Fixed an error in `Custom Function Node` port naming.
@@ -22,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - When you save a Shader Graph, all other Shader Graph windows no longer re-compile their preview Shaders.
 - Shader Graph UI now draws with correct styling for 2019.3.
 - When deleting edge connections to nodes with a preview error, input ports no longer draw in the wrong position.
+- Fixed an error involving deprecated components from VisualElements.
 
 ## [6.7.0-preview] - 2019-05-16
 ### Added