+++ com.unity.visualeffectgraph/CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Subgraphs
 - Move HDRP outputs to HDRP package + expose HDRP queue selection
 - Add exposure weight control for HDRP outputs
+- Shader macros for XR single-pass instancing
+- XR single-pass instancing support for indirect draws
 - Inverse trigonometric operators (atan, atan2, asin, acos)
 - Replaced Orient : Fixed rotation with new option Orient : Advanced